Added wg for proper shutdown
