+++ connector/failoverconnector/logs_test.go
@@ -134,197 +134,10 @@ func TestLogsWithFailoverError(t *testing.T) {
 	assert.EqualError(t, conn.ConsumeLogs(context.Background(), ld), "All provided pipelines return errors")
 }
 
-func TestLogsWithRecovery(t *testing.T) {
-	var sinkFirst, sinkSecond, sinkThird, sinkFourth consumertest.LogsSink
-	logsFirst := component.NewIDWithName(component.DataTypeLogs, "logs/first")
-	logsSecond := component.NewIDWithName(component.DataTypeLogs, "logs/second")
-	logsThird := component.NewIDWithName(component.DataTypeLogs, "logs/third")
-	logsFourth := component.NewIDWithName(component.DataTypeLogs, "logs/fourth")
-
-	cfg := &Config{
-		PipelinePriority: [][]component.ID{{logsFirst}, {logsSecond}, {logsThird}, {logsFourth}},
-		RetryInterval:    50 * time.Millisecond,
-		RetryGap:         10 * time.Millisecond,
-		MaxRetries:       10000,
-	}
-
-	router := connector.NewLogsRouter(map[component.ID]consumer.Logs{
-		logsFirst:  &sinkFirst,
-		logsSecond: &sinkSecond,
-		logsThird:  &sinkThird,
-		logsFourth: &sinkFourth,
-	})
-
-	conn, err := NewFactory().CreateLogsToLogs(context.Background(),
-		connectortest.NewNopCreateSettings(), cfg, router.(consumer.Logs))
-
-	require.NoError(t, err)
-
-	failoverConnector := conn.(*logsFailover)
-
-	lr := sampleLog()
-
-	defer func() {
-		assert.NoError(t, failoverConnector.Shutdown(context.Background()))
-	}()
-
-	t.Run("single failover recovery to primary consumer: level 2 -> 1", func(t *testing.T) {
-		defer func() {
-			resetLogsConsumers(failoverConnector, &sinkFirst, &sinkSecond, &sinkThird, &sinkFourth)
-		}()
-		failoverConnector.failover.ModifyConsumerAtIndex(0, consumertest.NewErr(errLogsConsumer))
-
-		require.NoError(t, conn.ConsumeLogs(context.Background(), lr))
-		idx := failoverConnector.failover.pS.TestStableIndex()
-		require.Equal(t, idx, 1)
-
-		failoverConnector.failover.ModifyConsumerAtIndex(0, &sinkFirst)
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 0, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-	})
-
-	t.Run("double failover recovery: level 3 -> 2 -> 1", func(t *testing.T) {
-		defer func() {
-			resetLogsConsumers(failoverConnector, &sinkFirst, &sinkSecond, &sinkThird, &sinkFourth)
-		}()
-		failoverConnector.failover.ModifyConsumerAtIndex(0, consumertest.NewErr(errLogsConsumer))
-		failoverConnector.failover.ModifyConsumerAtIndex(1, consumertest.NewErr(errLogsConsumer))
-
-		require.NoError(t, conn.ConsumeLogs(context.Background(), lr))
-		idx := failoverConnector.failover.pS.TestStableIndex()
-		require.Equal(t, idx, 2)
-
-		// Simulate recovery of exporter
-		failoverConnector.failover.ModifyConsumerAtIndex(1, &sinkSecond)
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 1, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-
-		failoverConnector.failover.ModifyConsumerAtIndex(0, &sinkFirst)
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 0, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-	})
-
-	t.Run("multiple failover recovery: level 3 -> 2 -> 4 -> 3 -> 1", func(t *testing.T) {
-		defer func() {
-			resetLogsConsumers(failoverConnector, &sinkFirst, &sinkSecond, &sinkThird, &sinkFourth)
-		}()
-		failoverConnector.failover.ModifyConsumerAtIndex(0, consumertest.NewErr(errLogsConsumer))
-		failoverConnector.failover.ModifyConsumerAtIndex(1, consumertest.NewErr(errLogsConsumer))
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 2, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-
-		// Simulate recovery of exporter
-		failoverConnector.failover.ModifyConsumerAtIndex(1, &sinkSecond)
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 1, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-
-		failoverConnector.failover.ModifyConsumerAtIndex(2, consumertest.NewErr(errLogsConsumer))
-		failoverConnector.failover.ModifyConsumerAtIndex(1, consumertest.NewErr(errLogsConsumer))
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 3, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-
-		failoverConnector.failover.ModifyConsumerAtIndex(2, &sinkThird)
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 2, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-
-		failoverConnector.failover.ModifyConsumerAtIndex(0, &sinkThird)
-
-		require.Eventually(t, func() bool {
-			return consumeLogsAndCheckStable(failoverConnector, 0, lr)
-		}, 3*time.Second, 5*time.Millisecond)
-	})
-}
-
-func TestLogsWithRecovery_MaxRetries(t *testing.T) {
-	var sinkFirst, sinkSecond, sinkThird, sinkFourth consumertest.LogsSink
-	logsFirst := component.NewIDWithName(component.DataTypeLogs, "logs/first")
-	logsSecond := component.NewIDWithName(component.DataTypeLogs, "logs/second")
-	logsThird := component.NewIDWithName(component.DataTypeLogs, "logs/third")
-	logsFourth := component.NewIDWithName(component.DataTypeLogs, "logs/fourth")
-
-	cfg := &Config{
-		PipelinePriority: [][]component.ID{{logsFirst}, {logsSecond}, {logsThird}, {logsFourth}},
-		RetryInterval:    50 * time.Millisecond,
-		RetryGap:         10 * time.Millisecond,
-		MaxRetries:       10000,
-	}
-
-	router := connector.NewLogsRouter(map[component.ID]consumer.Logs{
-		logsFirst:  &sinkFirst,
-		logsSecond: &sinkSecond,
-		logsThird:  &sinkThird,
-		logsFourth: &sinkFourth,
-	})
-
-	conn, err := NewFactory().CreateLogsToLogs(context.Background(),
-		connectortest.NewNopCreateSettings(), cfg, router.(consumer.Logs))
-
-	require.NoError(t, err)
-
-	failoverConnector := conn.(*logsFailover)
-
-	lr := sampleLog()
-
-	defer func() {
-		assert.NoError(t, failoverConnector.Shutdown(context.Background()))
-	}()
-
-	failoverConnector.failover.ModifyConsumerAtIndex(0, consumertest.NewErr(errLogsConsumer))
-	failoverConnector.failover.ModifyConsumerAtIndex(1, consumertest.NewErr(errLogsConsumer))
-
-	require.Eventually(t, func() bool {
-		return consumeLogsAndCheckStable(failoverConnector, 2, lr)
-	}, 3*time.Second, 5*time.Millisecond)
-
-	failoverConnector.failover.ModifyConsumerAtIndex(0, &sinkSecond)
-	failoverConnector.failover.ModifyConsumerAtIndex(1, &sinkSecond)
-
-	require.Eventually(t, func() bool {
-		return consumeLogsAndCheckStable(failoverConnector, 0, lr)
-	}, 3*time.Second, 5*time.Millisecond)
-
-	failoverConnector.failover.ModifyConsumerAtIndex(0, consumertest.NewErr(errLogsConsumer))
-	failoverConnector.failover.ModifyConsumerAtIndex(1, consumertest.NewErr(errLogsConsumer))
-	failoverConnector.failover.pS.SetRetryCountToMax(0)
-
-	require.Eventually(t, func() bool {
-		return consumeLogsAndCheckStable(failoverConnector, 2, lr)
-	}, 3*time.Second, 5*time.Millisecond)
-
-	failoverConnector.failover.ModifyConsumerAtIndex(0, &sinkSecond)
-	failoverConnector.failover.ModifyConsumerAtIndex(1, &sinkSecond)
-
-	require.Eventually(t, func() bool {
-		return consumeLogsAndCheckStable(failoverConnector, 1, lr)
-	}, 3*time.Second, 5*time.Millisecond)
-}
-
 func consumeLogsAndCheckStable(conn *logsFailover, idx int, lr plog.Logs) bool {
 	_ = conn.ConsumeLogs(context.Background(), lr)
 	stableIndex := conn.failover.pS.TestStableIndex()
 	return stableIndex == idx
-}
-
-func resetLogsConsumers(conn *logsFailover, consumers ...consumer.Logs) {
-	for i, sink := range consumers {
-
-		conn.failover.ModifyConsumerAtIndex(i, sink)
-	}
-	conn.failover.pS.TestSetStableIndex(0)
 }
 
 func sampleLog() plog.Logs {
