+++ connector/failoverconnector/metrics.go
@@ -21,8 +21,6 @@ type metricsFailover struct {
 	config   *Config
 	failover *failoverRouter[consumer.Metrics]
 	logger   *zap.Logger
-
-	done chan struct{}
 }
 
 func (f *metricsFailover) Capabilities() consumer.Capabilities {
@@ -62,19 +60,17 @@ func (f *metricsFailover) Shutdown(_ context.Context) error {
 	if f.failover != nil {
 		f.failover.Shutdown()
 	}
-	close(f.done)
 	return nil
 }
 
 func newMetricsToMetrics(set connector.CreateSettings, cfg component.Config, metrics consumer.Metrics) (connector.Metrics, error) {
 	config := cfg.(*Config)
 	mr, ok := metrics.(connector.MetricsRouterAndConsumer)
-	done := make(chan struct{})
 	if !ok {
 		return nil, errors.New("consumer is not of type MetricsRouter")
 	}
 
-	failover := newFailoverRouter[consumer.Metrics](mr.Consumer, config, done)
+	failover := newFailoverRouter[consumer.Metrics](mr.Consumer, config)
 	err := failover.registerConsumers()
 	if err != nil {
 		return nil, err
@@ -83,6 +79,5 @@ func newMetricsToMetrics(set connector.CreateSettings, cfg component.Config, met
 		config:   config,
 		failover: failover,
 		logger:   set.TelemetrySettings.Logger,
-		done:     done,
 	}, nil
 }
